Updated Journal.org Very weird critical error maybe due to checksum of svn_revisions

+++ Kameleon/debian7/execution/scripts/execo_script.py
@@ -19,46 +19,60 @@ class ExecoWorkload(Engine):
         if (not process.ok):
             logger.info("Error : {}".format(process.error_reason))
             logger.info("Spack stdout : {}".format(process.stdout))
+            return False
+        return True
 
     def run(self):
         f_csv = open(csv_file, 'r')
         csv_reader = csv.reader(f_csv, delimiter=',')
         header = CSVHeader(csv_reader.next())
 
-        # Go to the result folder before everything
-        os.chdir(self.result_dir)
-    
+        os.chdir(self.result_dir) # Go to result directory before everything
+
         while True:
             try:
                 row = csv_reader.next()
                 
                 logger.info("Starting experiment %s with Chameleon@%s and StarPU@%s..." % (row[header.name], row[header.chameleon_revision], row[header.starpu_revision]))
 
-                # STARPU INSTALLATION
-                spack_spec = 'chameleon@' + row[header.name] + '+starpu+fxt ^starpu@' + row[header.name] + '+fxt'
+                chameleon_name = row[header.name] + '_' + row[header.chameleon_revision]
+                starpu_name = row[header.name] + '_' + row[header.starpu_revision]
+                global_name = row[header.name] + '_c' + row[header.chameleon_revision] + '_s' + row[header.starpu_revision] 
+                
+                spack_spec = 'chameleon@' + chameleon_name + ' +starpu+fxt ^starpu@' + starpu_name + ' +fxt'
 
+                # FOLDER CREATION
+                folder = self.result_dir + '/' + global_name
+                os.mkdir(folder, 0764)
+
+                # STARPU INSTALLATION
                 logger.info("Starting StarPU installation...")
                 spack_process = Process('spack install -v' + ' ' + spack_spec)
 
-                spack_process.stdout_handlers.append(self.result_dir + '/' + 'StarPU_installation_' + row[header.name]) # create output file for StarPU installation
+                spack_process.stdout_handlers.append(folder + '/' + 'install_' + global_name) # create output file for StarPU installation
                 spack_process.start()
                 spack_process.wait()
 
                 logger.info("StarPU installation DONE...")
-                self.checkProcess(spack_process)
+                is_ok = self.checkProcess(spack_process)
                 spack_process.kill()
+
+                if (not is_ok):
+                    continue # stop this experiment
 
                 # STARPU DIRECTORY
                 logger.info("Searching and going to StarPU installation directory...")
 
                 starpu_location_process = Process('spack location -i' + ' ' + spack_spec).start()
                 starpu_location_process.wait()
-                self.checkProcess(starpu_location_process)
+                is_ok = self.checkProcess(starpu_location_process)
             
                 starpu_path = starpu_location_process.stdout.replace("\n", "") # remove end_of_line
                 starpu_cd = 'cd' + ' ' + starpu_path + '/lib/chameleon/'
-
                 starpu_location_process.kill()
+
+                if (not is_ok):
+                    continue # stop this experiment
                 
                 # RUNNING EXPERIMENT
                 logger.info("Starting StarPU experiment...")
@@ -69,13 +83,16 @@ class ExecoWorkload(Engine):
 
                 starpu_experiment_process = Process(starpu_cd + '\n' + starpu_experiment, shell=True)
                            
-                starpu_experiment_process.stdout_handlers.append(self.result_dir + '/' + 'StarPU_experiment_' + row[header.name]) # create output file for StarPU execution
+                starpu_experiment_process.stdout_handlers.append(folder + '/' + 'experiment_' + global_name) # create output file for StarPU execution
                 starpu_experiment_process.start()
                 starpu_experiment_process.wait()
 
                 logger.info("StarPU experiment DONE...")
-                self.checkProcess(starpu_experiment_process)        
+                is_ok = self.checkProcess(starpu_experiment_process)        
                 starpu_experiment_process.kill()
+
+                if (not is_ok):
+                    continue # stop this experiment
 
             except StopIteration:
                 break;
